Move Webpack scripts to scripts folder.
